+++ Connectors/Reqnroll.VisualStudio.ReqnrollConnector.Generic/ReqnrollProxies/Data/BindingScopeData.cs
@@ -6,4 +6,5 @@ public class BindingScopeData
     public string Tag { get; set; } // contains leading '@', e.g. '@mytag'
     public string FeatureTitle { get; set; }
     public string ScenarioTitle { get; set; }
+    public string Error { get; set; }
 }