+++ Reqnroll.VisualStudio/Discovery/Scope.cs
@@ -1,4 +1,4 @@
-using Reqnroll.VisualStudio.Discovery.TagExpressions;
+using Cucumber.TagExpressions;
 
 namespace Reqnroll.VisualStudio.Discovery;
 
@@ -7,6 +7,9 @@ public class Scope
     public ITagExpression? Tag { get; set; }
     public string? FeatureTitle { get; set; }
     public string? ScenarioTitle { get; set; }
+    public string? Error { get; set; }
+
+    public bool IsValid => Error == null;
 
     public override string ToString()
     {
@@ -20,6 +23,11 @@ public class Scope
         {
             result = result.Length > 0 ? result + ", " : result;
             result = $"{result}Scenario='{ScenarioTitle}'";
+        }
+        if (Error != null)
+        {
+            result = result.Length > 0 ? result + ", " : result;
+            result = $"{result}Error='{Error}'";
         }
         return result;
     }